add adapter - round #3
