+++ src/main/java/ru/ykul/Main.java
@@ -10,7 +10,7 @@ public class Main {
     private static int bagWeight = 50;
     private static int cost = 500;
     private static int discountStep = 5;
-    private static String inFileName = "discount_day";
+    private static String inFileName = "discount_day.txt";
     private static String outFileName = "orders.txt";
 
     public static void main(String[] args) {
