+++ src/main/java/ru/ykul/manager/OrderManager.java
@@ -4,7 +4,11 @@ import ru.ykul.model.Order;
 import ru.ykul.model.OrderReport;
 import ru.ykul.service.FileOrderService;
 import ru.ykul.service.OrderService;
+import ru.ykul.service.orderparsers.OrderParser;
+import ru.ykul.service.orderparsers.OrderParserByOctothorpe;
+import ru.ykul.service.orderparsers.OrderParserByPipe;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class OrderManager {
@@ -17,8 +21,17 @@ public class OrderManager {
     }
 
     public void writeOrderReport(String inFileName, String outFileName, double discount, double cost, double discountStep, int bagWeight) {
-        List<Order> orderlist = fileOrderService.read(inFileName);
-        OrderReport orderReport = orderService.createOrderReport(orderlist, discount, cost, discountStep, bagWeight);
+        String[] stringOrders = fileOrderService.read(inFileName);
+        List<Order> orders;
+        OrderParser orderParser;
+        if (inFileName.endsWith(".txt")) {
+            orderParser = new OrderParserByPipe();
+        }
+        else{
+            orderParser = new OrderParserByOctothorpe();
+        }
+        orders = orderParser.getOrderList(stringOrders);
+        OrderReport orderReport = orderService.createOrderReport(orders, discount, cost, discountStep, bagWeight);
         fileOrderService.write(orderReport, outFileName);
     }
 }
