+++ src/main/java/ru/ykul/service/orderparsers/OrderParser.java
@@ -1,0 +1,9 @@
+package ru.ykul.service.orderparsers;
+
+import ru.ykul.model.Order;
+
+import java.util.List;
+
+public interface OrderParser {
+    List<Order> getOrderList(String[] stringOrders);
+}
